+++ R/airnow_createMonitorObjects.R
@@ -5,6 +5,7 @@
 #' @param startdate desired start date (integer or character representing YYYYMMDD[HH])
 #' @param hours desired number of hours of data to assemble
 #' @param zeroMinimum logical specifying whether to convert negative values to zero
+#' @param addGoogleMeta logicial specifying wheter to use Google elevation and reverse geocoding services
 #' @return List where each element contains a \emph{ws_monitor} object for a unique parameter (e.g: "PM2.5", "NOX").
 #' @description This function uses the \link{airnow_downloadParseData} function 
 #' to download monthly dataframes of AirNow data and restructures that data into a format that is compatible
@@ -50,9 +51,10 @@
 airnow_createMonitorObjects <- function(parameters=NULL,
                                         startdate=strftime(lubridate::now(),"%Y%m%d",tz="GMT"),
                                         hours=24,
-                                        zeroMinimum=TRUE) {
+                                        zeroMinimum=TRUE,
+                                        addGoogleMeta=TRUE) {
   
-  metaList <- airnow_createMetaDataframes(parameters, 'AIRNOW')
+  metaList <- airnow_createMetaDataframes(parameters, 'AIRNOW', addGoogleMeta=addGoogleMeta)
   dataList <- airnow_createDataDataframes(parameters, startdate, hours)
   
   # Create empty list (no pre-allocation needed when lists are referenced by key instead of integer)