+++ R/epa_createMetaDataframe.R
@@ -1,5 +1,6 @@
 #' @keywords EPA
 #' @export
+#' @importFrom magrittr '%>%'
 #' @title Create Sites Metadata Dataframe
 #' @param tbl an EPA raw tibble after metadata enhancement
 #' @param pwfslDataIngestSource identifier for the source of monitoring data, e.g. \code{'EPA_hourly_88101_2016.zip'}
@@ -36,58 +37,64 @@ epa_createMetaDataframe <- function(tbl,
   
   logger.debug(paste0('Creating meta dataframe ...'))
   
-  # Subset tbl to include only site-specific columns
-  columns <- c('Site Num','Parameter Code','Latitude','Longitude',
-               'State Name','County Name','monitorID')
-  tbl <- dplyr::select(tbl, columns) %>% distinct()
+  # Subset tbl to contain only unique sites
+  siteColumns <- c('State Code', 'County Code', 'Site Num','Parameter Code',
+               'POC','Latitude','Longitude','State Name','County Name')
+  tbl <- dplyr::select(tbl, siteColumns) %>% distinct()
 
   # Our tibble now contains the following columns:
   #
   # > names(tbl)
-  # [1] "Site Num"       "Parameter Code" "Latitude"       "Longitude"     
-  # [5] "State Name"     "County Name"    "monitorID"         
+  #  [1] "State Code"          "County Code"         "Site Num"            "Parameter Code"     
+  #  [5] "POC"                 "Latitude"            "Longitude"           "Datum"              
+  #  [9] "Parameter Name"      "Date Local"          "Time Local"          "Date GMT"           
+  # [13] "Time GMT"            "Sample Measurement"  "Units of Measure"    "MDL"                
+  # [17] "Uncertainty"         "Qualifier"           "Method Type"         "Method Code"        
+  # [21] "Method Name"         "State Name"          "County Name"         "Date of Last Change"
   #
   # The PWFSLSmoke v1.0 data model contains the following parameters
   # 
-  #  [1] "monitorID"             "longitude"             "latitude"             
-  #  [4] "elevation"             "timezone"              "countryCode"          
-  #  [7] "stateCode"             "siteName"              "agencyName"           
-  # [10] "countyName"            "msaName"               "monitorType"          
-  # [13] "monitorInstrument"     "aqsID"                 "pwfslID"              
-  # [16] "pwfslDataIngestSource" "telemetryAggregator"   "telemetryUnitID"      
+  # > names(meta)
+  #  [1] "monitorID"             "longitude"             "latitude"              "elevation"            
+  #  [5] "timezone"              "countryCode"           "stateCode"             "siteName"             
+  #  [9] "agencyName"            "countyName"            "msaName"               "monitorType"          
+  # [13] "siteID"                "instrumentID"          "aqsID"                 "pwfslID"              
+  # [17] "pwfslDataIngestSource" "telemetryAggregator"   "telemetryUnitID"      
   
   # NOTE:  'meta' must be a dataframe because it has rownames which are deprecated in tibbles
   # Create empty dataframe
-  meta <- as.data.frame(matrix(nrow=nrow(tbl),ncol=18))
+  meta <- as.data.frame(matrix(nrow=nrow(tbl),ncol=19), stringsAsFactors=FALSE)
   
   colNames <- c("monitorID", "longitude", "latitude",
                 "elevation", "timezone", "countryCode",
                 "stateCode", "siteName", "agencyName",
                 "countyName", "msaName", "monitorType",
-                "monitorInstrument", "aqsID", "pwfslID",
+                "siteID", "instrumentID", "aqsID", "pwfslID",
                 "pwfslDataIngestSource", "telemetryAggregator", "telemetryUnitID")
   
   names(meta) <- colNames
   
   # Assign data where we have it
-  meta$monitorID <- tbl$monitorID
-  meta$longitude <- tbl$Longitude
-  meta$latitude <- tbl$Latitude
+  meta$longitude <- as.numeric(tbl$Longitude)
+  meta$latitude <- as.numeric(tbl$Latitude)
   meta$elevation <- as.numeric(NA)
   meta$timezone <- as.character(NA)
   meta$countryCode <- as.character(NA)
   meta$stateCode <- as.character(NA)
   meta$siteName <- as.character(NA)
-  meta$agencyName <- as.character(NA)
-  meta$countyName <- tbl$`County Name`
+  meta$countyName <- as.character(NA)
   meta$msaName <- as.character(NA)
-  meta$monitorType <- as.character(NA)
-  meta$monitorInstrument <- as.character(NA)
-  meta$aqsID <- tbl$monitorID
-  meta$pwfslID <- as.character(NA)
-  meta$pwfslDataIngestSource <- pwfslDataIngestSource
+  meta$agencyName <- as.character(NA)
+  meta$monitorType <- as.character(NA) # TODO:  Could figure this out by looking at 'Method Name' in original data
+  meta$siteID <- paste0(tbl$`State Code`,tbl$`County Code`,tbl$`Site Num`)
+  meta$instrumentID <- sprintf("%02d", as.numeric(tbl$POC))
+  meta$aqsID <- paste0(tbl$`State Code`,tbl$`County Code`,tbl$`Site Num`)
+  meta$pwfslID <- as.character(NA) # TODO:  This will be obtained from the "known_location" service
+  meta$pwfslDataIngestSource <- as.character(pwfslDataIngestSource)
   meta$telemetryAggregator <- as.character(NA)
   meta$telemetryUnitID <- as.character(NA)
+  
+  meta$monitorID <- paste(meta$siteID, meta$instrumentID, sep='_')
   
   # Assign rownames
   rownames(meta) <- meta$monitorID
@@ -97,39 +104,15 @@ epa_createMetaDataframe <- function(tbl,
   
   # TODO:  Could assign other spatial identifiers like EPARegion, etc.
   
-  # Add elevation, siteName and countyName
-  # NOTE:  Check and loop to see if we need to use addGoogleMetadata multiple 
-  # NOTE:  times due to google's per-query limit
-  loopCount <- nrow(meta) %/% 300
-  if ( nrow(meta) %% 300 > 0) {
-    loopCount <- loopCount + 1
-  }
-  
   if ( addGoogleMeta ) {
-    
-    logger.debug("Adding Google metadata ...")
-    metaList <- list()
-    for (i in 1:loopCount) {
-      startIndex <- (i-1) * 300 + 1
-      if (i != loopCount) {
-        endIndex <- i * 300
-      } else {
-        endIndex <- nrow(meta)
-      }
-      m <- addGoogleElevation(meta[startIndex:endIndex,], existingMeta=existingMeta)
-      m <- addGoogleAddress(m, existingMeta=existingMeta)
-      metaList[[i]] <- m
-    }
-    meta <- dplyr::bind_rows(metaList)
-    
+    # Add elevation, siteName and countyName
+    meta <- addGoogleElevation(meta, existingMeta=existingMeta)
+    meta <- addGoogleAddress(meta, existingMeta=existingMeta)
   }
-  
-  # Convert some columns to character even if they have all NA
-  characterColumns <- c('siteName','agencyName','countyName','msaName','monitorType',
-                        'monitorInstrument','aqsID')
-  for (colName in characterColumns) {
-    meta[[colName]] <- as.character(meta[[colName]])
-  }
+
+  # Restrict to North America
+  CANAMEX <- c('CA','US','MX')
+  airnowTbl <- dplyr::filter(airnowTbl, airnowTbl$countryCode %in% CANAMEX)
   
   logger.info("Created 'meta' dataframe with %d rows and %d columns", nrow(meta), ncol(meta))