+++ R/epa_createMonitorObject.R
@@ -2,6 +2,7 @@
 #' @export
 #' @title Download and Convert Hourly EPA Air Quality Data
 #' @param zipFile absolute path to monitoring data .zip file
+#' @param zeroMinimum logical specifying whether to convert negative values to zero
 #' @param addGoogleMeta logicial specifying wheter to use Google elevation and reverse geocoding services
 #' @description Convert EPA data into a \emph{ws_monitor} object, ready for use with all \code{monitor_~} functions.
 #' @note Before running this function you must first enable spatial data capabilities as in the example.
@@ -16,6 +17,7 @@
 #' }
 
 epa_createMonitorObject <- function(zipFile=NULL,
+                                    zeroMinimum=TRUE,
                                     addGoogleMeta=TRUE) {
 
   # Sanity checks
@@ -26,20 +28,6 @@ epa_createMonitorObject <- function(zipFile=NULL,
   
   tbl <- epa_parseData(zipFile)
       
-  # Create a column with the unique monitorID using fields 1:3 in the "file format description" above
-  tbl$monitorID <- paste0(tbl$`State Code`,tbl$`County Code`,tbl$`Site Num`)
-
-  # Create a column with the datetime
-  tbl$datetime <- lubridate::ymd_hms(paste0(tbl$`Date GMT`,' ',tbl$`Time GMT`,':00'))
-  
-  # check if MazamaSpatialUtils data are loaded 
-  if ( !exists('NaturalEarthAdm1') ) {
-    # NOTE:  Timezone, countryCode and stateCode information is mandatory for ws_monitor objects.
-    logger.error("MazamaSpatialUtils package was not properly initialized -- no Mazama metadata added")
-    stop(paste0("MazamaSpatialUtils package was not properly initialized.\n",
-                "Please run:  library(MazamaSpatialUtils); setSpatialDataDir('~/Data/Spatial'); loadSpatialData('NaturalEarthAdm1')"))
-  }
-  
   # Create 'meta' dataframe
   logger.debug("Creating 'meta' dataframe ...")
   meta <- epa_createMetaDataframe(tbl,
@@ -53,6 +41,12 @@ epa_createMonitorObject <- function(zipFile=NULL,
   # Create the 'ws_monitor' data list
   ws_monitor <- list(meta=meta, data=data)
   ws_monitor <- structure(ws_monitor, class = c("ws_monitor", "list"))
+
+  # Reset all negative values that made it through QC to zero
+  if ( zeroMinimum ) {
+    logger.debug("Reset negative valus to zero ...")
+    ws_monitor <- monitor_replaceData(ws_monitor, data < 0, 0)
+  }
   
   logger.debug(paste0('Finished creating ws_monitor object'))